Refactor and enhance user role handling and data management
- Removed unnecessary user role passing through `extra` in GoRoute.
- Enhanced `DataBaseMethouds` to handle user deletion based on role and phone number.
- Updated `BottomNavBar` to fetch user role using `CustomerDataProvider`.
- Improved `ProfileViewBody` to handle user role and phone number more effectively.
- Removed redundant user role passing in various authentication and navigation flows.
- Added `resetCustomerData` method in `CustomerDataProvider` to clear user data.
- Ensured consistent and secure user data management across the application.

+++ lib/Core/servers/data_base_methouds.dart
@@ -5,6 +5,9 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:food_delivery_app/featurs/home/Presentation/Manager/providers/customer_data_provider.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+import '../../featurs/auth/data/enums/user_role_enum.dart';
 
 class DataBaseMethouds {
   Future addUserDetails(Map<String, dynamic> userData, String uId) async {
@@ -22,17 +25,61 @@ class DataBaseMethouds {
     return FirebaseFirestore.instance.collection(name).snapshots();
   }
 
-
   Future<void> logOut() async {
     await FirebaseAuth.instance.signOut();
+    final SharedPreferences prefs = await SharedPreferences.getInstance();
+    await prefs.clear();
   }
 
-  Future<void> deleteUser() async {
-    await FirebaseFirestore.instance
-        .collection('users')
-        .doc(FirebaseAuth.instance.currentUser!.uid)
-        .delete();
-    await FirebaseAuth.instance.currentUser!.delete();
+  Future<void> deleteUser(
+      {required String userRole, required String phoneNumber}) async {
+    String collectionName = userRole == UserRoleEnum.user.getDisplayName
+        ? UserRoleEnum.user.getCollectionName
+        : UserRoleEnum.storeOwner.getCollectionName;
+    final docRef =
+        FirebaseFirestore.instance.collection(collectionName).doc(phoneNumber);
+    if (userRole == UserRoleEnum.user.getDisplayName) {
+      final subCollectoin = await docRef.collection('Purchases').get();
+      final subCollectoin2 = await docRef.collection('curt').get();
+    if (subCollectoin.docs.isNotEmpty || subCollectoin2.docs.isNotEmpty) {
+        if (subCollectoin.docs.isNotEmpty) {
+          for (var doc in subCollectoin.docs) {
+            await doc.reference.delete();
+          }
+        }
+
+        if (subCollectoin2.docs.isNotEmpty) {
+          for (var doc in subCollectoin2.docs) {
+            await doc.reference.delete();
+          }
+        }
+
+        await docRef.delete();
+      }
+    } else {
+      final subCollectoin = await docRef.collection('products').get();
+      final subCollectoin2 = await docRef.collection('orders').get();
+      if (subCollectoin.docs.isNotEmpty || subCollectoin2.docs.isNotEmpty) {
+        if (subCollectoin.docs.isNotEmpty) {
+          for (var doc in subCollectoin.docs) {
+            await doc.reference.delete();
+          }
+        }
+
+        if (subCollectoin2.docs.isNotEmpty) {
+          for (var doc in subCollectoin2.docs) {
+            await doc.reference.delete();
+          }
+        }
+
+        await docRef.delete();
+      }
+    }
+    if (FirebaseAuth.instance.currentUser != null) {
+      await FirebaseAuth.instance.currentUser!.delete();
+    } else {
+      await FirebaseAuth.instance.signOut();
+    }
   }
 
   Stream<QuerySnapshot> getCart(BuildContext context) async* {
@@ -57,13 +104,13 @@ class DataBaseMethouds {
     });
   }
 
-  Future deleteItemFromCurt(String id,BuildContext context) async {
+  Future deleteItemFromCurt(String id, BuildContext context) async {
     String? phoneNumber = context.read<CustomerDataProvider>().phoneNumber;
-      FirebaseFirestore.instance
+    FirebaseFirestore.instance
         .collection('users')
         .doc(phoneNumber)
-        .collection('curt').doc(id).delete();
-  
+        .collection('curt')
+        .doc(id)
+        .delete();
   }
-
 }
